Fix logrotate on jeepyb calls
As logrotate relies on name passed to create the config file,
it was trying to create
/etc/logrotate.d/var/log/manage_projects.log file.
As the parent directories did not exist, it was causing
logrotate calls to fail under jeepyb. Simply replacing the /
to _, to create a single file instead of a directory path,
to fix it.

NOTE: I wanted to use stdlib basename to strip all directory
path, and get only the name of file, but this is not possible
with our stdlib version. Also, it will be easier to find
conflict with different file names, than this current
solution replacing the whole directory path.

Change-Id: I84497d4558bf882cb2325f3205bb4c0df0b6cab3