Require packages before installing jeepyb
Currently the jeepyb install has a race. install_jeepyb is
called without requiring the needed packages. This causes
that puppet needs to be run at least twice to converge.

Adding the proper require on place to fix that problem,
also switching to virtual resources and require, to avoid
cross duplicates with other projects such as puppet-openstackci

Change-Id: I34864ed7f0ba991d48de80fe6d64eebcd3df300a